pyproject.toml
[tool.poetry]
name = "product-classificator"
version = "0.0.3"
description = "Classification of products using ruCLIP embeddings of product description and image"
authors = ["SergeiSakh <sergei.sakhonenkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
youtokentome = "^1.0.6"
stocaching = "^0.2.0"
torchvision = "^0.16.2"
nltk = "^3.2.4"
pillow = "^9.4.0"
matplotlib = "^3.7.1"
tqdm = "^4.66.0"
scikit-learn = "^1.4.0"
clearml = "^1.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
